+++ openspec/changes/archive/2026-05-21-a19-extract-members-bulk-csv-handlers/proposal.md
